lấy all đề và đổi sang examaction
