+++ frontend/src/actions/exams-action.js
@@ -37,6 +37,31 @@ export const getExams = async () => {
   }
 };
 
+export const ApprovedExamsList = async ({ page = 1, query }) => {
+  try {
+    const res = await instanceAPI.get("/exams/all", {
+      params: { page, query },
+    });
+
+    const allExams = res.data.data || [];
+    const approvedExams = allExams.filter((exam) => exam.status === "DA_DUYET");
+
+    return {
+      data: approvedExams,
+      totalPage: res.data.totalPage || 1,
+    };
+  } catch (error) {
+    console.error(
+      "Lỗi khi gọi API lấy đề thi:",
+      error?.response?.data || error
+    );
+    return {
+      data: [],
+      totalPage: 0,
+    };
+  }
+};
+
 export const getSignedExamFiles = async (examId) => {
   try {
     const res = await instanceAPI.get(`/exams/${examId}/files`);
