+++ frontend/src/actions/sign-action.js
@@ -25,30 +25,7 @@ export const getExams = async ({ page = 1, query }) => {
 };
 
 
-export const ApprovedExamsList = async ({ page = 1, query }) => {
-  try {
-    const res = await instanceAPI.get("/exams", {
-      params: { page, query },
-    });
 
-    const allExams = res.data.data || [];
-    const approvedExams = allExams.filter((exam) => exam.status === "DA_DUYET");
-
-    return {
-      data: approvedExams,
-      totalPage: res.data.totalPage || 1,
-    };
-  } catch (error) {
-    console.error(
-      "Lỗi khi gọi API lấy đề thi:",
-      error?.response?.data || error
-    );
-    return {
-      data: [],
-      totalPage: 0,
-    };
-  }
-};
 
 // export const uploadSignature = async (file, password) => {
 //   try {
